folly: ubsan: disable integer overflow tests in Histogram
Summary:
It would be nice to fix, but it's going to be easier to do when folly
support is for GCC>=5 which adds `__builtin_add_overflow`.

For now disable the check as the Histgram code handles overflow
correctly and has tests for these cases.

Fixes need to be done for both float and integer types as Histogram is used with both.

Reviewed By: meyering

Differential Revision: D3280260

fbshipit-source-id: 56e524c517566a4547346859be7770eda2acee96