add a new Locked class, to succeed Synchronized
Summary:
This adds a new Locked<T> class.  Locked<T> is very similar to Synchronized<T>,
but provides slightly different mechanisms to access the protected data, so
that preprocessor macros are not needed.

To illustrate the API differences, code using the older Synchronized API can be
changed from this:

  Synchronized<Foo> foo;
  foo->doStuff();  // grab and release the lock implicitly
  SYNCHRONIZED(lockedFoo, foo) {
    lockedFoo.doStuff();
    lockedFoo.otherStuff();
  }

to this:

  Locked<Foo> foo;
  foo.lock()->doStuff();
  foo.withLock([](auto& lockedFoo) {
    lockedFoo.doStuff();
    lockedFoo.otherStuff();
  });

or this:

  Locked<Foo> foo;
  foo.lock()->doStuff();
  {
    auto lockedFoo = foo.lock();
    lockedFoo->doStuff();
    lockedFoo->otherStuff();
  }

The SYNCHRONIZED macros had several minor annoying issues which could not be
fixed.  (Interference with break and continue statements when used inside
loops, and triggering false positive compiler warnings about missing return
statements.)

Besides avoiding the slightly troublesome macros, this new API makes the
locking more explicit.  This should help avoid some usage anti-patterns
mentioned in the Synchronized documentation where callers woudld acquire and
release the locks multiple times without realizing it.  The new APIs also
provide a cleaner mechanism for getting a read lock if you have a non-const
object.

Locked<T> also drops a few other APIs deemed unsafe (copy, move, and swap
operations that operate on the data but not the locks).

Test Plan: Ported all of the Synchronized tests to Locked.

Differential Revision: D3480966